backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-generativeai==0.8.0
googletrans==4.0.0rc1
langdetect==1.0.9
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.2
